New algorithm partially implemented
Now calculates TOTAL NGI/D instead of simply NGI/D. TNGI/D is multiplied
by the amount of seeds of that type that the user can afford. The
choosing seeds screen should have the TNGI/D dynamically change as
choices are made. Currently VERY buggy and not working as intended.